requirements/test.txt
# Test dependencies go here.
-r base.txt

coverage==4.4.1
flake8==3.9.2
django-test-plus==1.0.18
model-bakery==1.0.2
django-debug-toolbar==2.1
django-extensions==2.2.5

# pytest
pytest-django==3.1.2
django-coverage-plugin==1.8.0
